avance venta casi lista

+++ Clifford/src/CapaDTO/Producto.java
@@ -14,79 +14,47 @@ public class Producto {
     private int idProducto;
     private String nombreProducto;
     private String descripcionProducto;
-    private double precioUnitario;
+    private int precioUnitario;
     private int stockProducto;
 
-    /**
-     * @return the idProducto
-     */
     public int getIdProducto() {
         return idProducto;
     }
 
-    /**
-     * @param idProducto the idProducto to set
-     */
     public void setIdProducto(int idProducto) {
         this.idProducto = idProducto;
     }
 
-    /**
-     * @return the nombreProducto
-     */
     public String getNombreProducto() {
         return nombreProducto;
     }
 
-    /**
-     * @param nombreProducto the nombreProducto to set
-     */
     public void setNombreProducto(String nombreProducto) {
         this.nombreProducto = nombreProducto;
     }
 
-    /**
-     * @return the descripcionProducto
-     */
     public String getDescripcionProducto() {
         return descripcionProducto;
     }
 
-    /**
-     * @param descripcionProducto the descripcionProducto to set
-     */
     public void setDescripcionProducto(String descripcionProducto) {
         this.descripcionProducto = descripcionProducto;
     }
 
-    /**
-     * @return the precioUnitario
-     */
-    public double getPrecioUnitario() {
+    public int getPrecioUnitario() {
         return precioUnitario;
     }
 
-    /**
-     * @param precioUnitario the precioUnitario to set
-     */
-    public void setPrecioUnitario(double precioUnitario) {
+    public void setPrecioUnitario(int precioUnitario) {
         this.precioUnitario = precioUnitario;
     }
 
-    /**
-     * @return the stockProducto
-     */
     public int getStockProducto() {
         return stockProducto;
     }
 
-    /**
-     * @param stockProducto the stockProducto to set
-     */
     public void setStockProducto(int stockProducto) {
         this.stockProducto = stockProducto;
     }
-
-    
     
 }
